Replaces inserts + semantic hints with merges for bulk json loading
Our testing seems to show that while the IGNORE_ROW_ON_DUPKEY_INDEX
semantic hint successfully ignores attempts to insert duplicate
records when inserting one row at a time, it fails with insert
statements that include multiple rows. For the latter, this commit
removes inserts with semantic hints and replaces them with merge
statements that insert when not matched, and do nothing when matched,
effectively accomplishing the same thing, with good performance.