Improves use of Pure API metadata
We now include change metadata as an attribute of collection metadata,
to help ensure that we operations involving both changes and collections
use the correct, associated tables, i.e., those with matching API versions.

We temporarily remove the pureapi dependency, until we can update it to
use a newer version of attrs. We needed a newer version of attrs in this
package to implement some new collection metadata features.

We also shorten some names to make them more convenient and, we hope,
to eliminate confusion.